… src/main/java/jchess/el/TileExpression.java
… src/main/java/jchess/gamemode/square2p/Square2pPieces.java
… src/test/java/jchess/el/TileExpressionTest.java