"Verbessert" die Performance von TileExpressions um 10 Prozent (ich hatte auf mehr gehofft).
Testfall
```java
@Test
void test_neighborTiles() {
	Hex3PlayerGame hex3PlayerGame = new Hex3PlayerGame();
	hex3PlayerGame.start();
	Entity testTile = hex3PlayerGame.getEntityManager().getEntities().stream()
			.filter(entity -> entity.tile != null && entity.tile.position.x == 16 && entity.tile.position.y == 8)
			.findFirst().orElse(null);
	hex3PlayerGame.createPiece(testTile, PieceType.QUEEN, 0);

	Assertions.assertNotNull(testTile, "TestTile should have been found");
	Assertions.assertNotNull(testTile.piece, "TestTile should have received a piece");

        // "warm up" JVM
	for (int i = 0; i < 10_000; i++) {
		List<MoveIntention> validMoves = testTile.findValidMoves(true).toList();
	}

	long startTime = System.currentTimeMillis();
	for (int i = 0; i < 500; i++) {
		List<MoveIntention> validMoves = testTile.findValidMoves(true).toList();
	}
	long endTime = System.currentTimeMillis();
	logger.info("compute valid moves took {} ms", endTime - startTime);
}
```

Vorher: `6066 ms` (`12.132 ms` pro Aufruf)
Nachher: `5498 ms` (`10.996 ms` pro Aufruf)
Mit Parallelisierung: `4153 ms` (`8.306 ms` pro Aufruf)